# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SendR API",
    "description": "WeTransfer-like file sharing service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/request-code": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Request Code",
        "operationId": "request_code_api_auth_request_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Request Code Api Auth Request Code Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/verify-code": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify Code",
        "operationId": "verify_code_api_auth_verify_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
    "/api/auth/quota": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get Quota",
        "operationId": "get_quota_api_auth_quota_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotaResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CodeVerificationRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "create_account": {
            "type": "boolean",
            "title": "Create Account",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "email",
          "code"
        ],
        "title": "CodeVerificationRequest"
      },
...
      "EmailVerificationRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "EmailVerificationRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuotaResponse": {
        "properties": {
          "max_file_size_mb": {
            "type": "integer",
            "title": "Max File Size Mb"
          },
          "max_files_per_upload": {
            "type": "integer",
            "title": "Max Files Per Upload"
          },
          "weekly_uploads_limit": {
            "type": "integer",
            "title": "Weekly Uploads Limit"
          },
          "weekly_uploads_used": {
            "type": "integer",
            "title": "Weekly Uploads Used"
          },
          "weekly_uploads_remaining": {
            "type": "integer",
            "title": "Weekly Uploads Remaining"
          },
          "expiry_options_hours": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Options Hours"
          },
          "min_expiry_hours": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Expiry Hours"
          },
          "max_expiry_hours": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Expiry Hours"
          },
          "max_downloads_options": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Downloads Options"
          },
          "max_downloads_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Downloads Limit"
          },
          "max_passwords_per_upload": {
            "type": "integer",
            "title": "Max Passwords Per Upload",
            "default": 0
          },
          "max_emails_per_upload": {
            "type": "integer",
            "title": "Max Emails Per Upload",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "max_file_size_mb",
          "max_files_per_upload",
          "weekly_uploads_limit",
          "weekly_uploads_used",
          "weekly_uploads_remaining"
        ],
        "title": "QuotaResponse"
      },
...
      "TokenResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "token",
          "expires_at"
        ],
        "title": "TokenResponse"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "tier"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  }
}